+++ R/pkg-package.R
@@ -4,22 +4,21 @@
 #'
 #' @name wmisc-package
 #' @aliases wmisc-package wmisc
-#' @author Juergen Wilbert [aut, cre]
+#' @author Juergen Wilbert \[aut, cre\]
 #' @keywords package
 #' @import dplyr
 #' @import graphics
-#' @import ggplot2
+#' @importFrom utils read.csv citation packageVersion
+#' @importFrom ggplot2 ggplot geom_col geom_text aes scale_fill_brewer position_stack unit
+#'   theme_minimal ylab xlab theme element_blank coord_flip element_rect annotate 
 #' @importFrom gt gt html opt_table_lines tab_options opt_table_font
 #' @importFrom openxlsx write.xlsx
 #' @importFrom forcats fct_explicit_na
-#' @importFrom magrittr %>%
 #' @importFrom glue glue
 #' @importFrom lavaan fitMeasures
+#' @importFrom remotes package_deps install_github
 #' @importFrom psych scoreFast alpha describe fa fa.sort
 #' @importFrom nlme lme VarCorr gls
-#' @importFrom stats aggregate anova aov dnorm lm manova
-#' @importFrom kableExtra kable_classic kable_classic_2 row_spec footnote
-#'   kable_styling column_spec
-#' @importFrom knitr kable
-#' @importFrom stats median pchisq qf qnorm qt sd t.test
+#' @importFrom stats aggregate anova aov dnorm lm manova median pchisq qf pnorm
+#'   qnorm pt qt sd t.test var mad cor.test cor setNames coef residuals relevel chisq.test loadings quantile
 "_PACKAGE"